+++ CHANGELOG.md
@@ -45,7 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [cli] `--kc/--kitty-compress` 'kitty' style-specific option ([#51]).
 - [cli] `--query-timeout` command-line option ([3b658f3]).
 - [cli] `--itc/--iterm2-compress`, `--itjq/--iterm2-jpeg-quality` and `--itnrff/iterm2-no-read-from-file` style-specific command-line options ([#55]).
+- [cli] `-multi` command-line option.
 - [tui] Concurrent/Parallel frame rendering for TUI animations ([#42]).
+- [cli,tui] `no multi`, `query timeout` and `style` config options.
 - [lib,cli,tui] Support for the Kitty terminal graphics protocol ([#39]).
 - [lib,cli,tui] Automatic render style selection based on the detected terminal support ([#37]).